% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lksr.r
\name{loglikStatic}
\alias{loglikStatic}
\title{Convert a factor to numeric}
\usage{
loglikStatic(
  Y,
  w,
  omegaRho = 0,
  omegaVar = 0,
  omegaTrd = 0,
  density = "normal",
  df = NULL,
  result = "loglik",
  kernel = "epanechnikov"
)
}
\arguments{
\item{x}{A vector containing a factor with numeric levels}
}
\value{
The input factor made a numeric vector
}
\description{
Convert a factor with numeric levels to a non-factor
}
\examples{
x <- factor(c(3, 4, 9, 4, 9), levels=c(3,4,9))
fac2num(x)

}
