% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdsr.r
\name{loglikTVRhoCond}
\alias{loglikTVRhoCond}
\title{Convert a factor to numeric}
\usage{
loglikTVRhoCond(
  Y,
  w,
  omegaRho = 0,
  aRho = 0.01,
  bRho = 0.8,
  f1Rho = atanh(0.4),
  density = "normal",
  result = "loglik"
)
}
\arguments{
\item{x}{A vector containing a factor with numeric levels}
}
\value{
The input factor made a numeric vector
}
\description{
Convert a factor with numeric levels to a non-factor
}
\examples{
x <- factor(c(3, 4, 9, 4, 9), levels=c(3,4,9))
fac2num(x)

}
