% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumTrade}
\alias{sumTrade}
\title{Monthly total bilateral trade volume for the 29-economy panel}
\format{
A length-472 numeric vector.
}
\source{
See \link{W}.
}
\usage{
sumTrade
}
\description{
Monthly total bilateral trade volume for the 29-economy panel
}
\keyword{datasets}
