% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time}
\alias{time}
\title{POSIXct timestamps aligned with \link{W} and \link{Y}}
\format{
A length-472 \code{POSIXct} vector covering 1980-01 to 2019-04.
}
\source{
See \link{W}.
}
\usage{
time
}
\description{
POSIXct timestamps aligned with \link{W} and \link{Y}
}
\keyword{datasets}
