% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVSR-package.R
\docType{package}
\name{TVSR-package}
\alias{TVSR}
\alias{TVSR-package}
\title{TVSR: Codes for Time-Varying Spatial Regression Estimations}
\description{
Implementation of time-varying spatial regression models with two complementary approaches to modelling dynamics in the spatial-dependence coefficient. Local-Kernel Spatial Regression (LKSR) is a semi-parametric procedure that maximises a kernel-weighted local log-likelihood. Score-Driven Spatial Regression (SDSR) parameterises the spatial-dependence coefficient as a score-driven autoregressive recursion in the spirit of Blasques et al. (2016). Originally developed to study international monetary spillovers 1980-2020; see Peeters and Girard (2026) \url{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=6589023}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benjaminpeeters/TVSR}
  \item Report bugs at \url{https://github.com/benjaminpeeters/TVSR/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Peeters \email{benjaminpeeters@protonmail.com} (\href{https://orcid.org/0000-0002-5095-2021}{ORCID})

}
\keyword{internal}
