% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lksr.r
\name{crossValidLKSR}
\alias{crossValidLKSR}
\title{Cross-validaiton of Local-Kernel Spatial Regression}
\usage{
crossValidLKSR(Y, W, X = NULL, b, kernel = "epanechnikov", verbose = TRUE)
}
\arguments{
\item{x}{A vector containing a factor with numeric levels}
}
\value{
The input factor made a numeric vector
}
\description{
Convert a factor with numeric levels to a non-factor
}
\examples{
x <- factor(c(3, 4, 9, 4, 9), levels=c(3,4,9))
fac2num(x)

}
