% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lksr.r
\name{LKSR}
\alias{LKSR}
\title{Local-Kernel Spatial Regression}
\usage{
LKSR(Y, W, X = NULL, b = 1, kernel = "epanechnikov")
}
\arguments{
\item{x}{A vector containing a factor with numeric levels}
}
\value{
The input factor made a numeric vector
}
\description{
Convert a factor with numeric levels to a non-factor
Bandwidth: h = 2*b +1
}
\examples{
x <- factor(c(3, 4, 9, 4, 9), levels=c(3,4,9))
fac2num(x)

}
